Improve object oriented programming support.  Add curve support to timeline.
The context created for a tab now has the tab as its prototype, so keys that are not in  the context object,  its tab will be indexed firstly,  if not found, the class cocosContext will be tried.
g_t.timeline can use curves to control how the properties are animated.